% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\name{cell_from_extent}
\alias{cell_from_extent}
\title{Title}
\usage{
cell_from_extent(dimension, extent = NULL, x_extent)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}

\item{x_extent}{}
}
\description{
Title
}
