% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension.R
\name{cell_from_rowcol_combine}
\alias{cell_from_rowcol_combine}
\title{Title}
\usage{
cell_from_rowcol_combine(dimension, row, col)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{col}{}
}
\description{
Title
}
