% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\name{cell_from_xy}
\alias{cell_from_xy}
\title{Title}
\usage{
cell_from_xy(dimension, extent = NULL, xy)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}

\item{xy}{}
}
\description{
Title
}
