% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\name{cells}
\alias{cells}
\alias{cell_from_xy}
\alias{cell_from_extent}
\alias{extent_from_cell}
\alias{rowcol_from_cell}
\alias{xy_from_cell}
\alias{x_from_cell}
\alias{y_from_cell}
\alias{col_from_cell}
\alias{row_from_cell}
\alias{cell_from_row}
\alias{cell_from_col}
\alias{cell_from_row_col}
\alias{cell_from_rowcol_combine}
\title{Cells}
\usage{
cell_from_xy(dimension, extent = NULL, xy)

cell_from_extent(dimension, extent = NULL, x_extent)

extent_from_cell(dimension, extent = NULL, cell)

rowcol_from_cell(dimension, extent = NULL, cell)

xy_from_cell(dimension, extent = NULL, cell)

x_from_cell(dimension, extent = NULL, cell)

y_from_cell(dimension, extent = NULL, cell)

col_from_cell(dimension, cell)

row_from_cell(dimension, cell)

cell_from_row(dimension, row)

cell_from_col(dimension, col)

cell_from_row_col(dimension, row, col)

cell_from_rowcol_combine(dimension, row, col)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}

\item{xy}{matrix of coordinates}

\item{x_extent}{extent to find cells of}

\item{cell}{cells to find extent, or row,col, or xy of}

\item{row}{row to find cell of}

\item{col}{column to find cell of}
}
\value{
cell index

cells of extent

extent of cells

row,col of cells

xy from cells

x of cells

y of cells

col of cells

row of cells

cell of rows

cell of cols

cell of row,col

cell of row,col combined
}
\description{
Functions that work with cells.
}
\examples{
cell_from_xy(c(10, 5), extent = c(0, 10, 0, 5), cbind(5, 4))
cell_from_extent(c(10, 5), c(0, 10, 0, 5), c(6, 7, 2, 3))
extent_from_cell(c(10, 5), c(0, 10, 0, 5), c(4, 5))
rowcol_from_cell(c(10, 5), c(0, 10, 0, 5), 3:5)
xy_from_cell(c(10, 5), c(0, 10, 0, 5), 4:6)
x_from_cell(c(10, 5), c(0, 10, 0, 5), 4:7)
y_from_cell(c(10, 5), c(0, 10, 0, 5), 4:7)
col_from_cell(c(10, 5),  4:7)
row_from_cell(c(10, 5),  4:7)
cell_from_row(c(10, 5), 4:7)
cell_from_col(c(10, 5), 4:7)
cell_from_row_col(c(10, 5), 1:4, 4:7)
cell_from_rowcol_combine(c(10, 5), 1:4, 4:7)
}
