% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{col_from_x}
\alias{col_from_x}
\title{Title}
\usage{
col_from_x(dimension, extent = NULL, x)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}

\item{x}{}
}
\description{
Title
}
