% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coords}
\alias{coords}
\title{Title}
\usage{
coords(dimension, extent = NULL)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}
}
\description{
Title
}
