% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_extent}
\alias{draw_extent}
\title{Draw extent}
\usage{
draw_extent(show = TRUE, ...)
}
\arguments{
\item{...}{arguments pass to \code{\link[graphics:rect]{graphics::rect()}}}

\item{plot}{the drawn extent}
}
\description{
Draw an extent with two clicks
}
