% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\name{extent_from_cell}
\alias{extent_from_cell}
\title{Title}
\usage{
extent_from_cell(dimension, extent = NULL, cells)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}

\item{cells}{}
}
\description{
Title
}
