% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dims.R
\name{fit_dims}
\alias{fit_dims}
\title{Aspect ratio of dimension conflated with bbox}
\usage{
fit_dims(size = 1024L, wh = c(size, size))
}
\arguments{
\item{size}{seed dimension size}

\item{wh}{distance across dimension span/s}
}
\value{
description dimension \code{c(ncol, nrow)}
}
\description{
Generate an appropriate dimension (shape, ncol,nrow) from an
input width(height). If heignt not specified we have a square.
}
\examples{
fit_dims(256, c(10, 20))
fit_dims(1024, c(102723, 1e5))
}
