% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_xyz.R
\name{from_xyz}
\alias{from_xyz}
\title{Derive a grid from XYZ points}
\usage{
from_xyz(xyz, digits = 5)
}
\arguments{
\item{xyz}{set of points xy or xyz (matrix or data frame)}

\item{digits}{argument passed to \code{\link[=round]{round()}}}
}
\value{
list with elements 'dimension', 'extent'
}
\description{
This function is very liberal, it simply finds unique x values and unique y values,
sorts them and finds the minimum difference between the points, then checks that rounded
ratio of differences to this minimum is 1.
}
\details{
The points can be the full grid set, a partial set, or a superset of the grid. The resolution will simply be
the smallest actual difference found. (Zero is not possible because we \code{sort(unique(...))}).

The z-column if present is ignored.
}
\examples{
from_xyz(vaster_long(c(10, 5), c(0, 10, 0, 5)))
}
