% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{grid}
\alias{grid}
\alias{n_cell}
\alias{x_res}
\alias{y_res}
\alias{n_row}
\alias{n_col}
\alias{xlim}
\alias{ylim}
\alias{x_min}
\alias{x_max}
\alias{y_min}
\alias{y_max}
\title{Grid}
\usage{
n_cell(dimension)

x_res(dimension, extent = NULL)

y_res(dimension, extent = NULL)

n_row(dimension)

n_col(dimension)

xlim(dimension, extent = NULL)

ylim(dimension, extent = NULL)

x_min(dimension, extent = NULL)

x_max(dimension, extent = NULL)

y_min(dimension, extent = NULL)

y_max(dimension, extent = NULL)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}
}
\value{
number of cells

x resolution (width of cell)

y resolution (height of cell)

number of rows

number of cols

x extent (corner to corner)

y extent (corner to corner)

x minimum (left edge)

x maximum (right edge)

y minimum (bottom edge)

ymaximum (top edge)
}
\description{
Basic grid tools, cell, resolution, dimension, extent.
}
\examples{
n_cell(c(10, 5))
x_res(c(10, 5), c(0, 10, 0, 5))
y_res(c(10, 5), c(0, 10, 0, 5))
n_row(c(10, 5))
n_col(c(10, 5))
xlim(c(10, 5), c(0, 10, 0, 5))
ylim(c(10, 5), c(0, 10, 0, 5))
x_min(c(10, 5), c(0, 10, 0, 5))
x_max(c(10, 5), c(0, 10, 0, 5))
y_min(c(10, 5), c(0, 10, 0, 5))
y_max(c(10, 5), c(0, 10, 0, 5))
}
