% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_extent}
\alias{plot_extent}
\title{Plot an extent}
\usage{
plot_extent(x, ..., asp = 1, add = FALSE, border = "black")
}
\arguments{
\item{x}{extent xmin,xmax,ymin,ymax}

\item{...}{arguments passed to \code{\link[graphics:rect]{graphics::rect()}}}

\item{asp}{aspect ratio (1 by default)}

\item{add}{add to plot or initiated one, \code{FALSE} by default}

\item{border}{colour of lines of extent}
}
\value{
nothing, used for plot side effect
}
\description{
Plot an extent
}
\examples{
plot_extent(c(-180, 180, -90, 90))
plot_extent(c(100, 150, -60, -30), add = TRUE, border = "firebrick")
}
