% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{rasterio_idx}
\alias{rasterio_idx}
\alias{raster_sfio}
\title{The sf RasterIO is the RasterIO window in a list format used by the sf package, it contains the same
information, and is created by \code{\link[=raster_sfio]{raster_sfio()}}.}
\usage{
rasterio_idx(dimension, extent)

raster_sfio(dimension, fact = 1, resample = "Nearest")
}
\arguments{
\item{dimension}{ncols, nrows}

\item{extent}{this is ignored}

\item{fact}{a resizing factor}

\item{resample}{resample algorithm for GDAL RasterIO}
}
\value{
RasterIO window vector 'c(x0, y0, nx0, ny0, nx, y)' see Details
}
\description{
The sf RasterIO is the RasterIO window in a list format used by the sf package, it contains the same
information, and is created by \code{\link[=raster_sfio]{raster_sfio()}}.
}
\examples{
rasterio_idx(dim(volcano))
}
