% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{row_from_y}
\alias{row_from_y}
\title{Title}
\usage{
row_from_y(dimension, extent = NULL, y)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}

\item{y}{}
}
\description{
Title
}
