% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaster.R
\name{snap_extent}
\alias{snap_extent}
\alias{buffer_extent}
\title{Snap extent to resolution (buffer extent)}
\usage{
snap_extent(x, res)

buffer_extent(x, res)
}
\arguments{
\item{x}{extent (xmin, xmax, ymin, ymax)}

\item{res}{resolution (a grain to align to)}
}
\description{
Whole grain buffers.
}
\examples{
snap_extent(sort(rnorm(4)), 0.01)
}
