% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{vaster_boundary}
\alias{vaster_boundary}
\title{Grid boundary in native resolution}
\usage{
vaster_boundary(dimension, extent = NULL)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}
}
\description{
currently only return centre coords
}
\examples{
vaster_boundary(c(3, 4))
}
