% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{vaster_boundary_cell}
\alias{vaster_boundary_cell}
\title{Grid boundary cell index}
\usage{
vaster_boundary_cell(dimension)
}
\arguments{
\item{dimension}{integer ncol, nrow}
}
\value{
a matrix of xy coordinates
}
\description{
This is for indexing coordinate arrays to get their values (the cell index of the outer row and columns).
}
\details{
The orientation is the same as for \code{\link[=vaster_boundary]{vaster_boundary()}}.
}
\examples{
vaster_boundary_cell(c(3, 4))
cell <- vaster_boundary_cell(c(3, 4))
plot(vaster_boundary(c(3, 4)))
text(xy <- xy_from_cell(c(3, 4), cell = vaster_boundary_cell(c(3, 4))), lab = cell)
lines(xy)
}
