% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaster.R
\name{vcrop}
\alias{vcrop}
\title{Virtual grid modification}
\usage{
vcrop(x, dimension, extent = NULL, ..., snap = "out")
}
\arguments{
\item{x}{extent of candidate grid (vector of xmin,xmax,ymin,ymax)}

\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}

\item{...}{ignored}

\item{snap}{one of "out" (default), "near", or "in"}
}
\description{
To modify a grid is to align an extent to the grid origin.  Modification includes to reduce or extend the area covered,
in either dimension.  This implies a new extent, snapped
to the grain of the origin grid and a new size (dimension in x,y).
}
\details{
This works for any grid, the input extent can be within the original, an extension of the original, or
completely non-intersecting the original grid.
}
\examples{
## any arbitrary extent
x <- c(sort(runif(2, -180, 180)), sort(runif(2, -90, 90)))
print(x)
vcrop(x,  c(360, 180), c(-180, 180, -90, 90))
}
