% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{x_corner}
\alias{x_corner}
\title{Title}
\usage{
x_corner(dimension, extent = NULL)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}
}
\description{
Title
}
