% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\name{x_from_cell}
\alias{x_from_cell}
\title{Title}
\usage{
x_from_cell(dimension, extent = NULL, cell)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}

\item{cell}{}
}
\description{
Title
}
