% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{x_from_col}
\alias{x_from_col}
\title{Title}
\usage{
x_from_col(dimension, extent = NULL, col)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}

\item{col}{}
}
\description{
Title
}
