% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{y_from_row}
\alias{y_from_row}
\title{Title}
\usage{
y_from_row(dimension, extent = NULL, row)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}
}
\description{
Title
}
