% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencies.R
\name{adjacencies}
\alias{adjacencies}
\alias{bl}
\alias{tl}
\alias{br}
\alias{tr}
\alias{la}
\alias{ta}
\alias{ra}
\alias{ba}
\title{Adjacency, for use in creating area based meshes}
\usage{
bl(x)

tl(x)

br(x)

tr(x)

la(x)

ta(x)

ra(x)

ba(x)
}
\value{
matrix, padded by one row and one column relative to input
}
\description{
Functions 'bottom left', 'top left', 'bottom right', and 'top right' named by
their initials, provide very low level relative positional structures for use in
raster logic. These are used to traverse the divide left by area-based rasters which are
inherently a discrete value across a finite element. If we want that element as part of a
continuous surface we need to find local relative values for its corners. Used in
quadmesh and anglr packages, and useful for calculating neighbourhood values.
}
\details{
bl, tl, br, and tr originally lived in affinity
}
\examples{
(m <- matrix(1:12, 3))
tl(m)
tr(m)
bl(m)
br(m)
tl(br(m))
image0(tl(br(m)))
text0(tl(br(m)))

## this gives neighbours in adjacent positions
m <- matrix(1:12, ncol = 3, byrow = TRUE)

matrix(c(t(la(m)), t(ta(m)), t(ra(m)), t(ba(m))), ncol = 4)

## this gives neighbours in all 8 adjacent and diagonal positions
ximage(matrix(rowMeans(matrix(c(t(la(m)), t(ta(m)), t(ra(m)), t(ba(m)), t(bal(m)), t(tal(m)), t(bar(m)), t(tar(m))),  ncol = 8), na.rm = TRUE), 4, byrow = TRUE))
}
