% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaster.R
\name{align_extent}
\alias{align_extent}
\title{Title}
\usage{
align_extent(x, dimension, extent = NULL, snap = c("out", "near", "in"))
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}

\item{snap}{}
}
\description{
Title
}
