% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaster.R
\name{intersect_extent}
\alias{intersect_extent}
\title{Title}
\usage{
intersect_extent(x, dimension, extent = NULL)
}
\arguments{
\item{x}{extent to intersect}

\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}
}
\description{
Title
}
