% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaster.R
\name{intersect_extent}
\alias{intersect_extent}
\title{Intersect extent}
\usage{
intersect_extent(x, dimension, extent = NULL)
}
\arguments{
\item{x}{extent to intersect}

\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}
}
\value{
extent
}
\description{
Return the overlapping extent.
}
\examples{
intersect_extent(c(0.5, 2.3, 1.2, 5), c(10, 5), c(0, 10, 0, 5))
}
