% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaster-package.R
\docType{package}
\name{vaster-package}
\alias{vaster}
\alias{vaster-package}
\title{vaster: Tools for Raster Grid Logic}
\description{
Provides raster grid logic, operations that describe a discretized rectangular domain and don't require access to materialized data. Grids are arrays with dimension and extent, and many operations are functions of just the dimension 'nrows', 'ncols' or a combination of the dimension and the extent 'xmin', 'xmax', 'ymin', 'ymax'. Here we provide direct access to this logic without need for connection to any materialized data or formats. Grid logic includes functions that relate the cell index to row and column, or row and column to cell index, row, column or cell index to position.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/vaster}
  \item \url{https://hypertidy.github.io/vaster/}
  \item Report bugs at \url{https://github.com/hypertidy/vaster/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

Other contributors:
\itemize{
  \item Robert Hijmans (Wrote original versions of abstract cell operations in raster package) [contributor]
}

}
\keyword{internal}
