% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{vaster_boundary}
\alias{vaster_boundary}
\title{Grid boundary in native resolution}
\usage{
vaster_boundary(dimension, extent = NULL)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}
}
\value{
matrix of xy coordinates
}
\description{
Generates the coordinates around a grid boundary, with a coordinate for
each grid cell corner.
}
\details{
The orientation is starts along the bottom and goes counter-clockwise.
}
\examples{
vaster_boundary(c(3, 4))

plot(vaster_boundary(c(36, 18), c(-180, 180, -90, 90)))
}
