% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{vaster_listxyz}
\alias{vaster_listxyz}
\title{Image trad form}
\usage{
vaster_listxyz(dimension, extent = NULL, data = NULL)
}
\arguments{
\item{dimension}{integer ncol, nrow}

\item{extent}{numeric extent xmin,xmax,ymin,ymax}

\item{data}{}
}
\description{
Image trad form
}
\examples{
## see https://gist.github.com/mdsumner/b844766f28910a3f87dc2c8a398a3a13
}
