% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{vaster_long}
\alias{vaster_long}
\title{Convert to long form (XYZ\link{,Z2,etc})}
\usage{
vaster_long(extent, dimension, data = NULL, raster_order = TRUE)
}
\arguments{
\item{raster_order}{}
}
\description{
Matrix of x, y\link{, data} in raster order, use 'raster_order = FALSE' for traditional R matrix 'x\link{i}, y\link{i}' order
}
